requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
djangorestframework==3.14.0
joblib==1.2.0
numpy==1.24.1
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
scikit-learn==0.24.2
django-cors-headers = "*"
